+++ macros/src/parser/state_machine.rs
@@ -1,13 +1,14 @@
 use super::transition::{StateTransition, StateTransitions};
-use syn::{braced, parse, spanned::Spanned, token, Ident, Token, Type};
+use syn::{braced, bracketed, parse, spanned::Spanned, token, Ident, Token, Type};
 
 #[derive(Debug)]
 pub struct StateMachine {
     pub temporary_context_type: Option<Type>,
     pub custom_guard_error: bool,
-    pub impl_display_states: bool,
-    pub impl_display_events: bool,
     pub transitions: Vec<StateTransition>,
+    pub name: Option<Ident>,
+    pub derive_states: Vec<Ident>,
+    pub derive_events: Vec<Ident>,
 }
 
 impl StateMachine {
@@ -15,9 +16,10 @@ impl StateMachine {
         StateMachine {
             temporary_context_type: None,
             custom_guard_error: false,
-            impl_display_states: false,
-            impl_display_events: false,
             transitions: Vec::new(),
+            name: None,
+            derive_states: Vec::new(),
+            derive_events: Vec::new(),
         }
     }
 
@@ -78,20 +80,6 @@ impl parse::Parse for StateMachine {
                     }
 
                 }
-                "impl_display_states" => {
-                    input.parse::<Token![:]>()?;
-                    let b: syn::LitBool = input.parse()?;
-                    if b.value {
-                        statemachine.impl_display_states = true
-                    }
-                }
-                "impl_display_events" => {
-                    input.parse::<Token![:]>()?;
-                    let b: syn::LitBool = input.parse()?;
-                    if b.value {
-                        statemachine.impl_display_events = true
-                    }
-                }
                 "temporary_context" => {
                     input.parse::<Token![:]>()?;
                     let temporary_context_type: Type = input.parse()?;
@@ -116,10 +104,46 @@ impl parse::Parse for StateMachine {
                     statemachine.temporary_context_type = Some(temporary_context_type);
 
                 }
+                "name" =>{
+                    input.parse::<Token![:]>()?;
+                    statemachine.name = Some(input.parse::<Ident>()?);
+                },
+                "derive_states" => {
+                    input.parse::<Token![:]>()?;
+                    if input.peek(token::Bracket) {
+                        let content;
+                        bracketed!(content in input);
+                        loop{
+                            if content.is_empty() {
+                                break;
+                            };
+                            let trait_ =  content.parse::<Ident>()?;
+                            statemachine.derive_states.push(trait_);
+                            if content.parse::<Token![,]>().is_err() {
+                                break;
+                            };
+                        }
+                    }
+                },
+                "derive_events" => {
+                    input.parse::<Token![:]>()?;
+                    let content;
+                    bracketed!(content in input);
+                    loop{
+                        if content.is_empty() {
+                            break;
+                        };
+                        let trait_ =  content.parse::<Ident>()?;
+                        statemachine.derive_events.push(trait_);
+                        if content.parse::<Token![,]>().is_err() {
+                            break;
+                        };
+                    }
+                },
                 keyword => {
                     return Err(parse::Error::new(
                         input.span(),
-                        format!("Unknown keyword {}. Support keywords: [\"transitions\", \"temporary_context\", \"custom_guard_error\"]", keyword)
+                        format!("Unknown keyword {}. Support keywords: [\"name\", \"transitions\", \"temporary_context\", \"custom_guard_error\", \"derive_states\", \"derive_events\"]", keyword)
                     ))
                 }
             }